pyproject.toml
[tool.poetry]
name = "multi-modular-image-search"
version = "0.1.0"
description = ""
authors = ["Priyank Gupta <priyankgupta2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
python-multipart = "^0.0.17"
pillow = "^11.0.0"
requests = "^2.32.3"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
clickhouse-connect = "^0.8.6"
transformers = "^4.46.3"
torch = "^2.5.1"
tqdm = "^4.67.0"
chromadb = "^0.5.20"
aiosqlite = "^0.20.0"
boto3 = "^1.35.65"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
